# #2782 - Faster symmetric_interval_hull of singletons