Closes #2782.

```julia
julia> S = rand(Singleton, dim=1000);

julia> @time symmetric_interval_hull(S);  # master
  0.002510 seconds (5 allocations: 31.781 KiB)

julia> @time symmetric_interval_hull(S);  # this branch
  0.000012 seconds (3 allocations: 15.906 KiB)
```
---
```julia
julia> @time symmetric_interval_hull(M*S);  # master
  2.491252 seconds (2.01 k allocations: 15.534 MiB)

julia> @time symmetric_interval_hull(concretize(M*S));  # proposal from #2782
  0.007192 seconds (7 allocations: 39.750 KiB)

julia> @time symmetric_interval_hull(M*S);  # this branch
  0.000738 seconds (5 allocations: 23.875 KiB)
```
---
```julia
julia> @time symmetric_interval_hull(S+S);  # master
  0.005102 seconds (6 allocations: 31.812 KiB)

julia> @time symmetric_interval_hull(concretize(S+S));  # idea from #2782
  0.002465 seconds (7 allocations: 39.750 KiB)

julia> @time symmetric_interval_hull(S+S);  # this branch
  0.000018 seconds (5 allocations: 23.875 KiB)

```